Extract platform capabilities shell bootstrap module
